systhreads library: merge all C code in a single file st_stubs.c
Now that the Pthreads API is used on all supported platforms, even
Windows, we no longer need to "functorize" over the API using .h files
conditionally included.  Instead, we can just put all the C code in
the st_stubs.c.  There's only one function containing a `#ifdef WIN32`.
