--- otherlibs/systhreads/st_win32.h
@@ -1,27 +1,0 @@
-/**************************************************************************/
-/*                                                                        */
-/*                                 OCaml                                  */
-/*                                                                        */
-/*          Xavier Leroy and Damien Doligez, INRIA Rocquencourt           */
-/*                                                                        */
-/*   Copyright 2009 Institut National de Recherche en Informatique et     */
-/*     en Automatique.                                                    */
-/*                                                                        */
-/*   All rights reserved.  This file is distributed under the terms of    */
-/*   the GNU Lesser General Public License version 2.1, with the          */
-/*   special exception on linking described in the file LICENSE.          */
-/*                                                                        */
-/**************************************************************************/
-
-/* Win32 implementation of the "st" interface */
-
-#undef _WIN32_WINNT
-#define _WIN32_WINNT 0x0400
-#include <windows.h>
-
-Caml_inline void st_msleep(int msec)
-{
-  Sleep(msec);
-}
-
-#include "st_pthreads.h"
